requirements.txt
Flask==3.0.3
gunicorn==22.0.0
twilio==9.3.5
python-dotenv==1.0.1
black==25.1.0

google-api-python-client==2.141.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1